Issue #3219325 by fgm, klelostec: support missing file info on functions.
- Also fix KV changes in D9.
- Also fix various D9.2 changes
- PHP 8 support

+++ modules/mongodb_watchdog/src/Controller/ControllerBase.php
@@ -157,7 +157,7 @@ abstract class ControllerBase extends CoreControllerBase {
    * @return int
    *   The actual index of the page to display.
    */
-  public static function getPage(int $count, int $requestedPage, int $height): int {
+  public static function getPage(float $count, int $requestedPage, int $height): int {
     if ($requestedPage <= 0) {
       return 0;
     }
